python_alembic/requirements.txt
alembic
pymysql
SQLAlchemy
cryptography
python-dotenv
# psycopg2
# psycopg2-binary